py/requirements.txt
# Flask Backend Requirements
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0

# YouTube Processing
youtube-transcript-api==0.6.1

# AI/ML Libraries
anthropic==0.39.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Text Processing
nltk==3.8.1

# Environment Management
python-dotenv==1.0.0

# API Documentation (optional)
flask-swagger-ui==4.11.1